feat: Update theme handling and tmux plugins
This commit updates the way themes are handled in alacritty and tmux.
The toggle-theme.toml file has been removed and replaced with a direct
reference to the theme file in alacritty.toml. The dark-theme.toml and
light-theme.toml files have also been deleted.

In the install.sh script, the copy command for alacritty themes has been
replaced with a link command. This change ensures that the themes are
linked correctly.

The toggle-themes.sh script in tmux has been updated to directly copy
the desired theme file instead of modifying the toggle-theme.toml file.
This change simplifies the theme switching process.

A new tmux plugin, tmux-monokai-pro, has been added to the tmux.conf
file. This plugin provides a new theme option for tmux.