feat: Use symbolic links for alacritty themes
This commit changes the way we handle alacritty themes. Instead of
copying the theme files, we now create symbolic links. This makes it
easier to update the themes as we can just update the source file and
the changes will be reflected in the linked file. Also, the redundant
variable `alacritty_config_file` is removed and its usage is replaced
with a direct path.